Fixed Several Issues after Pulling from Repo

+++ frontend/src/Pages/Candidates.jsx
@@ -229,7 +229,7 @@ const Candidates = () => {
   };
 
   const handleDelete = async (id) => {
-    {
+    if (window.confirm('Are you sure you want to delete this candidate?')) {
       try {
         await deleteCandidate(id);
         fetchData();
